client/requirements.txt
Pillow==10.0.1
pyperclip==1.8.2
requests==2.31.0